[HB] Fix pedantic warnings

+++ src/hb-buffer.h
@@ -56,13 +56,13 @@ typedef struct _hb_glyph_position_t {
   hb_position_t  x_advance;
   hb_position_t  y_advance;
   /* XXX these should all be replaced by "uint32_t internal" */
-  unsigned short new_advance :1;	/* if set, the advance width values are
+  uint32_t       new_advance :1;	/* if set, the advance width values are
 					   absolute, i.e., they won't be
 					   added to the original glyph's value
 					   but rather replace them */
-  unsigned short back : 15;		/* number of glyphs to go back
+  uint32_t       back : 15;		/* number of glyphs to go back
 					   for drawing current glyph */
-  short          cursive_chain : 16;	/* character to which this connects,
+  int32_t        cursive_chain : 16;	/* character to which this connects,
 					   may be positive or negative; used
 					   only internally */
 } hb_glyph_position_t;
